Remove calls to functions which don't exist
Round two of testing for Uplift - `undefined method `to_slug' Did you mean?  to_s`.

As far as I understand it, a mailing name already *is* a string, certainly this is true in Uplift's configuration, so I'm suggesting we remove `to_slug.transliterate`. Ideally, I would also prefer to remove the split on the `-`, which seems quite particular to how you name your mailings, but I'll leave that discussion for another time!